% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extractEveryNth}
\alias{extractEveryNth}
\title{Generate Time Frame Intervals for API Requests}
\usage{
extractEveryNth(n_months = 6, by_time = "15 mins", n = 3900)
}
\arguments{
\item{n_months}{Number of months of data to retrieve.}

\item{by_time}{Time interval for data extraction (e.g., '15 mins').}

\item{n}{Maximum number of data points per request.}
}
\value{
A list of start and stop timestamps.
}
\description{
This helper function generates start and stop timestamps to bypass the data point limitation of the Gain Capital API.
}
\examples{
\dontrun{
intervals <- extractEveryNth(n_months = 6, by_time = '15 mins', n = 3900)
}
}
