% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCapiClient.R
\docType{class}
\name{GCapiClient}
\alias{GCapiClient}
\title{GCapiClient}
\format{
An \code{\link{R6Class}} generator object
}
\description{
GCapiClient

GCapiClient
}
\details{
This R6 class provides methods to interact with the Gain Capital API V1 for trading on Forex.com.
}
\examples{
\dontrun{
gc_client <- GCapiClient$new("your_username", "your_password", "your_appkey")
account_info <- gc_client$get_account_info()
market_info <- gc_client$get_market_info("EUR/USD")
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rest_url}}{The base URL for the Gain Capital API.}

\item{\code{session_id}}{The session ID for the current session.}

\item{\code{username}}{The username for the current session.}

\item{\code{session}}{The CURL handle for the current session.}

\item{\code{trading_account_id}}{The trading account ID for the current session.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GCapiClient-new}{\code{GCapiClient$new()}}
\item \href{#method-GCapiClient-get_account_info}{\code{GCapiClient$get_account_info()}}
\item \href{#method-GCapiClient-get_market_info}{\code{GCapiClient$get_market_info()}}
\item \href{#method-GCapiClient-get_prices}{\code{GCapiClient$get_prices()}}
\item \href{#method-GCapiClient-get_ohlc}{\code{GCapiClient$get_ohlc()}}
\item \href{#method-GCapiClient-trade_order}{\code{GCapiClient$trade_order()}}
\item \href{#method-GCapiClient-list_open_positions}{\code{GCapiClient$list_open_positions()}}
\item \href{#method-GCapiClient-list_active_orders}{\code{GCapiClient$list_active_orders()}}
\item \href{#method-GCapiClient-close_all_trades}{\code{GCapiClient$close_all_trades()}}
\item \href{#method-GCapiClient-close_all_trades_new}{\code{GCapiClient$close_all_trades_new()}}
\item \href{#method-GCapiClient-get_trade_history}{\code{GCapiClient$get_trade_history()}}
\item \href{#method-GCapiClient-get_long_series}{\code{GCapiClient$get_long_series()}}
\item \href{#method-GCapiClient-clone}{\code{GCapiClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCapiClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-GCapiClient-new}{}}}
\subsection{Method \code{new()}}{
Initialize the GCapiClient object and create a session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCapiClient$new(username, password, appkey, proxies = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{The username for the Gain Capital API.}

\item{\code{password}}{The password for the Gain Capital API.}

\item{\code{appkey}}{The application key for the Gain Capital API.}

\item{\code{proxies}}{Optional. A list of proxy settings.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCapiClient-get_account_info"></a>}}
\if{latex}{\out{\hypertarget{method-GCapiClient-get_account_info}{}}}
\subsection{Method \code{get_account_info()}}{
Retrieve account information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCapiClient$get_account_info(get = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{get}}{Optional. Specific information to retrieve.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing account information or a specific value if `get` is provided.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCapiClient-get_market_info"></a>}}
\if{latex}{\out{\hypertarget{method-GCapiClient-get_market_info}{}}}
\subsection{Method \code{get_market_info()}}{
Retrieve market information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCapiClient$get_market_info(market_name, get = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{market_name}}{The name of the market to retrieve information for.}

\item{\code{get}}{Optional. Specific information to retrieve.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing market information or a specific value if `get` is provided.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCapiClient-get_prices"></a>}}
\if{latex}{\out{\hypertarget{method-GCapiClient-get_prices}{}}}
\subsection{Method \code{get_prices()}}{
Retrieve price data for a market.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCapiClient$get_prices(
  market_id,
  num_ticks,
  from_ts,
  to_ts,
  price_type = "MID"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{market_id}}{The market ID to retrieve prices for.}

\item{\code{num_ticks}}{The number of price ticks to retrieve.}

\item{\code{from_ts}}{The start timestamp in UTC.}

\item{\code{to_ts}}{The end timestamp in UTC.}

\item{\code{price_type}}{The type of price to retrieve ("MID", "BID", "ASK").}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table containing the price data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCapiClient-get_ohlc"></a>}}
\if{latex}{\out{\hypertarget{method-GCapiClient-get_ohlc}{}}}
\subsection{Method \code{get_ohlc()}}{
Retrieve OHLC data for a market.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCapiClient$get_ohlc(
  market_id = NULL,
  num_ticks = NULL,
  interval = "HOUR",
  span = 1,
  from_ts,
  to_ts
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{market_id}}{The market ID to retrieve OHLC data for.}

\item{\code{num_ticks}}{The number of ticks to retrieve.}

\item{\code{interval}}{The interval for OHLC data ("MINUTE", "HOUR", "DAY").}

\item{\code{span}}{The span of the interval.}

\item{\code{from_ts}}{The start timestamp in UTC.}

\item{\code{to_ts}}{The end timestamp in UTC.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table containing the OHLC data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCapiClient-trade_order"></a>}}
\if{latex}{\out{\hypertarget{method-GCapiClient-trade_order}{}}}
\subsection{Method \code{trade_order()}}{
Place a trade order.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCapiClient$trade_order(
  quantity,
  offer_price,
  bid_price,
  direction,
  trading_acc_id,
  market_id,
  market_name,
  stop_loss = NULL,
  take_profit = NULL,
  trigger_price = NULL,
  close = FALSE,
  order_id = NULL,
  tolerance = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantity}}{The quantity to trade.}

\item{\code{offer_price}}{The offer price for the trade.}

\item{\code{bid_price}}{The bid price for the trade.}

\item{\code{direction}}{The direction of the trade ("buy" or "sell").}

\item{\code{trading_acc_id}}{The trading account ID.}

\item{\code{market_id}}{The market ID.}

\item{\code{market_name}}{The market name.}

\item{\code{stop_loss}}{Optional. The stop loss price.}

\item{\code{take_profit}}{Optional. The take profit price.}

\item{\code{trigger_price}}{Optional. The trigger price.}

\item{\code{close}}{Optional. If TRUE, close the trade.}

\item{\code{order_id}}{Optional. The order ID.}

\item{\code{tolerance}}{Optional. The price tolerance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table containing the order details.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCapiClient-list_open_positions"></a>}}
\if{latex}{\out{\hypertarget{method-GCapiClient-list_open_positions}{}}}
\subsection{Method \code{list_open_positions()}}{
List open positions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCapiClient$list_open_positions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.table containing the open positions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCapiClient-list_active_orders"></a>}}
\if{latex}{\out{\hypertarget{method-GCapiClient-list_active_orders}{}}}
\subsection{Method \code{list_active_orders()}}{
List active orders.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCapiClient$list_active_orders()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing the active orders.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCapiClient-close_all_trades"></a>}}
\if{latex}{\out{\hypertarget{method-GCapiClient-close_all_trades}{}}}
\subsection{Method \code{close_all_trades()}}{
Close all trades.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCapiClient$close_all_trades(tolerance)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tolerance}}{The price tolerance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing the responses for each closed trade.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCapiClient-close_all_trades_new"></a>}}
\if{latex}{\out{\hypertarget{method-GCapiClient-close_all_trades_new}{}}}
\subsection{Method \code{close_all_trades_new()}}{
Close all trades using provided open positions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCapiClient$close_all_trades_new(open_positions, tolerance)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{open_positions}}{A list of open positions.}

\item{\code{tolerance}}{The price tolerance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing the responses for each closed trade.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCapiClient-get_trade_history"></a>}}
\if{latex}{\out{\hypertarget{method-GCapiClient-get_trade_history}{}}}
\subsection{Method \code{get_trade_history()}}{
Retrieve trade history.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCapiClient$get_trade_history(
  trading_account_id = self$trading_account_id,
  from = NULL,
  max_results = 100
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trading_account_id}}{The trading account ID.}

\item{\code{from}}{Optional. The start timestamp for trade history.}

\item{\code{max_results}}{The maximum number of results to return.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table containing the trade history.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCapiClient-get_long_series"></a>}}
\if{latex}{\out{\hypertarget{method-GCapiClient-get_long_series}{}}}
\subsection{Method \code{get_long_series()}}{
Retrieve long series data by bypassing the API limitation of max 4000 data points.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCapiClient$get_long_series(
  market_id,
  n_months = 6,
  by_time = "15 mins",
  n = 3900,
  interval = "MINUTE",
  span = 15
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{market_id}}{The market ID to retrieve the long series data for.}

\item{\code{n_months}}{Number of months of data to retrieve.}

\item{\code{by_time}}{Time interval for data extraction.}

\item{\code{n}}{Maximum number of data points per request.}

\item{\code{interval}}{Time interval for OHLC data ("MINUTE", "HOUR", "DAY").}

\item{\code{span}}{Span of the interval.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table containing the long series data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GCapiClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GCapiClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GCapiClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
