% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_instruction_status_description}
\alias{get_instruction_status_description}
\title{Get Instruction Status Description}
\usage{
get_instruction_status_description(status_code)
}
\arguments{
\item{status_code}{The status code to look up.}
}
\value{
A character string with the description of the status code.
}
\description{
Retrieve the description of an instruction status code.
}
