% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_instruction_status_reason_description}
\alias{get_instruction_status_reason_description}
\title{Get Instruction Status Reason Description}
\usage{
get_instruction_status_reason_description(reason_code)
}
\arguments{
\item{reason_code}{The reason code to look up.}
}
\value{
A character string with the description of the reason code.
}
\description{
Retrieve the description of an instruction status reason code.
}
