% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_order_status_descriptions}
\alias{get_order_status_descriptions}
\title{Get Order Status Descriptions}
\usage{
get_order_status_descriptions(status_code)
}
\arguments{
\item{status_code}{The status code to look up.}
}
\value{
A character string with the description of the status code.
}
\description{
Retrieve the description of an order status code.
}
