% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_order_status_reason_descriptions}
\alias{get_order_status_reason_descriptions}
\title{Get Order Status Reason Descriptions}
\usage{
get_order_status_reason_descriptions(reason_code)
}
\arguments{
\item{reason_code}{The reason code to look up.}
}
\value{
A character string with the description of the reason code.
}
\description{
Retrieve the description of an order status reason code.
}
