% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupTabs.R
\docType{data}
\name{instruction_status_reason_descriptions}
\alias{instruction_status_reason_descriptions}
\title{Instruction Status Reason Descriptions}
\format{
A named list with instruction status reason codes as keys and descriptions as values.
}
\usage{
instruction_status_reason_descriptions
}
\description{
A lookup table for instruction status reason codes and their corresponding descriptions.
}
\examples{
\dontrun{
instruction_status_reason_descriptions[["1"]]
}
}
\keyword{datasets}
