% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupTabs.R
\docType{data}
\name{order_action_type_descriptions}
\alias{order_action_type_descriptions}
\title{Order Action Type Descriptions}
\format{
A named list with Order action type codes as keys and descriptions as values.
}
\usage{
order_action_type_descriptions
}
\description{
A lookup table for Order action type codes and their corresponding descriptions.
}
\examples{
\dontrun{
order_action_type_descriptions[["1"]]
}

}
\keyword{datasets}
