% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupTabs.R
\docType{data}
\name{order_status_reason_descriptions}
\alias{order_status_reason_descriptions}
\title{Order Status Reason Descriptions}
\format{
A named list with order status reason codes as keys and descriptions as values.
}
\usage{
order_status_reason_descriptions
}
\description{
A lookup table for order status reason codes and their corresponding descriptions.
}
\examples{
\dontrun{
order_status_reason_descriptions[["1"]]
}
}
\keyword{datasets}
